backend_fastapi/requirements.txt
scikit-learn==1.0.2
protobuf==3.20.*
tensorflow-cpu==2.8.0
matplotlib
numpy
opencv-python
joblib
scipy
python-math
pandas
seaborn
pathlib
streamlit
requests
altair